Add files via upload
Uploaded HTML, CSS and image files of Paradise Crossing - An Unofficial Website for Animal Crossing Gameplay and More: 
For the WEBSITE FROM SCRATCH section of the assignment, I made a website using the basic HTML and CSS codes (ie. using tables and block, etc) and breakpoints for a responsive design. Please check the layout on the following breakpoints: 
@media all and (min-width: 600px) 
@media all and (min-width: 768px) 
@media all and (min-width: 992px) and (max-width: 1199) 
@media all and (min-width: 1200px)